TIL: 250619

+++ ch4-spring-db2/ch4-labs/src/main/java/com/example/ch4labs/domain/Review.java
@@ -19,5 +19,5 @@ public class Review {
     private String bookTitle;
     @Column(name = "book_author")
     private String bookAuthor;
-    private byte rating;
+    private Integer rating;
 }
