+++ ch4-spring-db2/ch4-labs/src/main/java/com/example/ch4labs/dto/ReviewCreateRequest.java
@@ -10,7 +10,7 @@ public class ReviewCreateRequest {
     private String author;
     private String bookTitle;
     private String bookAuthor;
-    private byte rating;
+    private Integer rating;
 
     public Review toDomain() {
         Review review = new Review();
