+++ ch4-spring-db2/ch4-labs/src/main/java/com/example/ch4labs/dto/ReviewResponse.java
@@ -13,7 +13,7 @@ public class ReviewResponse {
     private String author;
     private String bookTitle;
     private String bookAuthor;
-    private byte rating;
+    private Integer rating;
 
     public static ReviewResponse from(Review review) {
         return new ReviewResponse(
