+++ ch4-spring-db2/ch4-labs/src/main/java/com/example/ch4labs/dto/ReviewUpdateRequest.java
@@ -9,5 +9,5 @@ public class ReviewUpdateRequest {
     private String author;
     private String bookTitle;
     private String bookAuthor;
-    private byte rating;
+    private Integer rating;
 }
